Seperate Intl and Domestic classes for OnePay
